autotest: Start using Iperf in Perf tests
Add logic to network_WiFi_Perf to allow the control files to specify
whether the perf test should use netperf or iperf. This takes the form
of a use_iperf param which is parsed as an additional_param.

This change adds additional control files to the WiFi_Perf testing suite
to run some tests using the Iperf tool instead of the Netperf tool.
These test runs are supplemental to the existing Netperf test runs, so
we won't lose any Netperf coverage for the time being. Once these test
have run for a while and appear stable, we should start running Iperf by
default for all tests. The tests ht40_iperf and vht80_iperf were added
to allow a comparison to the current netperf test runs. Once these tests
are shown to be stable, they should be removed, and the existing tests
should be changed to use iperf.
(http://issuetracker.google.com/198693658)

This change also makes some small changes to the netperf testing logic
to make it more consistent with the iperf logic. For example, the
NetperfSession object is now defined per test run, rather than using a
single session for all the runs. This change also removes the warmup
logic that would run before perf tests, since I couldn't find any
evidence that it was necessary or useful.

One caveat, bidirectional tests in iperf are unreliable, so netperf is
still used for bidirectional tests, regardless of which tool was
requested.

Reference doc: go/cros-wifi-iperf-design

BUG=b:195574472
TEST=test_that ${DUT} suite:wifi_perf

Change-Id: I0084e12aadffebe66353209ad9109379a5c7272c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3133630
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Abhishek Kumar <kuabhs@chromium.org>
Tested-by: Kevin Lund <kglund@google.com>
Commit-Queue: Kevin Lund <kglund@google.com>