bluetooth: Power on the peer at the end of advmon interleave test
We have a mechanism to remove the DUT from the peer at the end of the
test. However, since the peer is powered off in the middle of the test,
that operation will fail.

To fix that, power on the peer at the end of the corresponding test.

BUG=b:199524297
TEST=verify DUT is removed after running
     bluetooth_AdapterAdvMonitor.advmon_interleaved_scan_tests

Change-Id: Ifb3ab2aad67af591ea5e5b2bdf7eefb3567fca53
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3154076
Reviewed-by: Yun-Hao Chung <howardchung@chromium.org>
Tested-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Archie Pusaka <apusaka@chromium.org>