autotest: Update wificell docs with link to partnersite
The autotest docs include a file 'wificell.md' which describes WiFi
testing cells that are used for WiFi testing. These docs provide a
general overview of wificells and specific considerations that apply for
autotest, but they do not include a full tutorial for setting up a
wificell yourself. This change adds a link to such a tutorial which
exists on the partnersite.

BUG=None
TEST=~/trunk/docs/scripts/preview_docs wificell.md

Change-Id: Ib4745c605e95cdc32beaeaadbf6647a9a60ce6b4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3160130
Reviewed-by: Hsin-chen Chuang <chharry@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>