dhcp: Make moving to namespace optional
With upgraded dhcpcd (which is properly unicasting requests during
renewal) it is important to run the DHCP server in a separate namespace
in order to do not change routing on DUT.  So by default new DHCP
related tests are run with one part of virtual_ethernet_pair moved to
'autotest' namespace.

This commit makes this optional (set it to None) for the case of other
tests that are based on dhcp_test_base but do not really need this
additional namespace.  It also allows to change the name used.

BUG=b:172213156
TEST=Run any DHCP related test

Change-Id: If855e34d160ba482f3bc18c8490f8cabd3a1705b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3168775
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>