VirtualEthernetPair: Better connectivity in namespace
When one end of ethernet pair is moved to another namespace you will
have connection from it to the DUT (once DHCP negotiation finishes).
However the "nested" world will have no connection to the outside.

No test has been identified so far as impacted but this might be needed,
so in order to fix this let's add some routing, and configure iptables
to allow forwarding from the "nested" interface and configure NAT.  With
this changes (once the address on peer interface is configured) you
should have the same connectivity as DUT (traffic from "nested" is
routed back via ethernet pair and 'eth0' on DUT acts as a gateway).

BUG=None
TEST=Testing is manual:
  - either execute these commands in DUT's shell or
  - modify some DHCP test (e.g. DhcpStaticIP) to run ping (in the
    'autotest' namespace!) and check that it is working

Change-Id: Iedfd4c341aa50f0333ecf6dff705cd87a52c6fc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3168776
Tested-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Andrzej Ostruszka <andrzejo@google.com>
Reviewed-by: Jun Yu <junyuu@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>