autotest: Replace ch011 with ch006 AttenuatedPerf
The test is only testing ch001 and ch011 in the 2.4GHz spectrum and they
are both edge channel and susceptible to the Band-edge power reduction.

Adding a middle channel (ch006 one of the mostly used channel based on
UMA metrics [1]) to the test to cover one of the channels that are not
afftected by that and removing channel 11.

BUG=b:183463625
TEST=ran
test_that --board=$BOARD $DUT network_WiFi_AttenuatedPerf.ht40_ch006

[1] https://uma.googleplex.com/histograms?sid=a52c61500a97f586578ef33fc4231823#Platform.WiFiStatusAfterForcedPCIRescan

Change-Id: Ica09d63944c54d7d2fe9ee73c83716266ed68c81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3163306
Tested-by: Arowa Suliman <arowa@chromium.org>
Reviewed-by: Harpreet Grewal <harpreet@chromium.org>
Commit-Queue: Arowa Suliman <arowa@chromium.org>