autotest: use security_config in WifiCellTestBase
Within configure_and_connect_to_ap(), WifiCellTestBase does not
currently utilize the security_config member of hostap_config when
initializing the WiFi connection. This means that when a security config
is required, the caller must bypass configure_and_connect_to_ap() and
call the required methods directly.

This is causing some performance tests to fail, since the devices cannot
connect without the proper security config.
http://stainless/search?suite=%5Ewifi_perf%24&exclude_cts=true&exclude_non_release=true&view=matrix&row=test&col=build&first_date=2021-09-15&last_date=2021-09-21

This change updates configure_and_connect_to_ap() in
WiFiCellTestBase to use the security_config value of the hostap_config
object. This should be safe, since if a hostap_config object has
security_config value, the connection will fail if that value is not
used during connection initialization. That is, there should no be any
cases where the security_config is present in the hostap_config object,
but was deliberately omitted from the connection setup, since this would
make the connection fail. This is backed up by the run of wifi_matfunc,
which passed 100% with this change.

BUG=b:200064611
TEST=test_that $DUT suite:wifi_matfunc
Total PASS: 24/24 (100%)

Change-Id: I3311f265e0b7e14eb6afca272afe6bf515f9828a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3170346
Commit-Queue: Kevin Lund <kglund@google.com>
Tested-by: Kevin Lund <kglund@google.com>
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>