autotest update login_LoginSuccess to handle gobject update
Because of the gobject module being updated for PY3 migrations, and the
AU tests being specifically non-ssp, there is no good way to know if the
DUT has gobject, or gi. So we have to account for both.

BUG=b:200722511
TEST=this

Change-Id: I28eb9ff3812d8d2267adb37b6169b0622238fc83
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3174357
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: David Haddock <dhaddock@chromium.org>
Reviewed-by: David Haddock <dhaddock@chromium.org>