cheest_CTS(_P/R): Add waiver runnners to the "cts-hardware" suite.
Those manually written test jobs need to be run on the
variant build suite, when some of the test cases in the
hardware-dependent test suite is under a waiver. Those
jobs will run the DEV version CTS that contains a fix
CL for the waived test. Submitting the pass results from
DEV version is the requirement for the failures to be
waived.

BUG=b:198384191
TEST=none

Change-Id: Ica967e62f88300d113d02436635b25c412b56e6d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3172959
Tested-by: Kazuhiro Inaba <kinaba@chromium.org>
Auto-Submit: Kazuhiro Inaba <kinaba@chromium.org>
Commit-Queue: Jaydeep Mehta <jaydeepmehta@google.com>
Reviewed-by: Jaydeep Mehta <jaydeepmehta@google.com>