autottest: Refine the logic of generating chart host name
In crrev/3156283, we assume all the host name of the camerabox DUT
should be in format: "FOO.cros". However, in the real case, it seems to
be "FOO" instead.

Therefore, this CL refines the logic to make it similar to how
DUT.CompanionDeviceHostname() does to append the suffix to the DUT host
name.

BUG=b:200890798
TEST=test_that -b {board} [CameraBoxHost] tast.camerabox-facing-front

Change-Id: Ic9e8f5fda2a94b553f674174f57428475b68c756
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3177454
Tested-by: Wei Lee <wtlee@chromium.org>
Auto-Submit: Wei Lee <wtlee@chromium.org>
Commit-Queue: Shuhei Takahashi <nya@chromium.org>
Reviewed-by: Shuhei Takahashi <nya@chromium.org>