Collect crossystem log for tests run in 'fast' mode
The lab team relies on the output of crossystem command for parsing. It
was found that when tests ran with the --fast flag, collection of
crossystem data was not occurring.

This change adds in the collection of crossystem logs as an after test
hook.

Note: This does not add in per-boot collection of crossystem like
non-fast mode run tests collect.

BUG=b:194804915
TEST=Ran stub_PassServer test with --fast and without the argument, and
confirmed the proper logs collected.

Change-Id: I55894bc7cb0bab1fc428911eb4a109568acaf831
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3194129
Tested-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Jesse McGuire <jessemcguire@google.com>