arcappcompat: Update maximum runtime for appcompat_release suite
It is observed that the appcompat_release suite couldn’t complete
the runs due to the global timeout reached error. Currently
appcompat_release tests are set to a global timeout of 120 minutes.
As new test cases were added to the suite,  120 minutes isn't enough
for running the test suite.  So increasing the timeout will fix the
issue.

BUG=b:201338436
TEST=test_that --board=eve <ip-address> tast.appcompat_release

Change-Id: Id9de951e82012868c4179635d568ebb47622588e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3194859
Tested-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Commit-Queue: Meenakshi Thiyagarajan <mthiyagarajan@google.com>
Reviewed-by: Seewai Fu <seewaifu@google.com>
Reviewed-by: Renuga Nanjappan <rnanjappan@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>