[autotest]: Verify that servo does not have issue with initialize dut_controller
Introduce a verifier to check whether the dut controller missing fault
is set to 'on' when servod is operting in recovery mode.

When the servod is operating on recovery mode, it will not be able to
interact with the DUT if the duct controller is missing. To gracefully
handle such situations, we need to check this using a verifier.

Cherry-picked from https://crrev.com/c/3174598

BUG=b:188106718
TEST=unittest

Change-Id: I8d02350e83d7d3c559587c002d53cd9b2f60d245
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/labpack/+/3174598
Tested-by: Vineet Joshi <vkjoshi@google.com>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Commit-Queue: Vineet Joshi <vkjoshi@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3195373