autotest: Add pvs-display suite
This contains the display tests that can be run by partners via moblab.
This suite will be used by partners to verify functionality of the display
tests using moblab. The suite is not intended for critical testing in the
near term.

For more information on PVS see go/cros-pvs.

BUG=b:200203512
TEST=Run by using 'test_that -b dedede 10.0.0.196 --autotest_dir
~/trunk/src/third_party/autotest/files/ suite:pvs-display

Change-Id: I3c74a4d1ad1bf3f7d92ded5d8e3dc9c2cf99f209
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3200326
Reviewed-by: Mike Wiitala <mwiitala@google.com>
Reviewed-by: Jack Neus <jackneus@google.com>
Reviewed-by: Jared Loucks <jaredloucks@google.com>
Reviewed-by: Tatsuhisa Yamaguchi <yamaguchi@chromium.org>
Tested-by: Eduardo Valdes <valdese@chromium.org>
Commit-Queue: Eduardo Valdes <valdese@chromium.org>