servo_host: expand MCU_EXTRACTOR regex
For some systems, the cr50 console is tagged as 'PD/Cr50' in the logs.
The current extractor fails to recognize that as a cr50 logline and will
skip them.
This change fixes the regex to account for the '/' as well.
This introduces one issue - the MCU name is used to generate a filename,
and we should avoid '/' in the filename. The change simply replaces all
instances of '/' with '.' in the mcu name.

BUG=b:198777617
TEST=None

Change-Id: I4839dd442b8297c0614007522b7b62f42d47abc4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3200575
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Xianuo Wang <xianuowang@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>