Bluetooth: refresh adapter properties upon bluetoothd start
adapter properties are stored in bluetooth_facade at the very
beginning of the test. If bluetoothd stopped in the previous test run,
the stored adapter properties are wrong. This adds code when we start
bluetoothd before the main test.

BUG=b:201348088
TEST=stop bluetoothd on DUT and run any bluetooth test. Verify
"Trying to remove None" is not in the chameleond log.

Change-Id: Ifed164928165cdbbc576f36ee360076cc5118339
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3212282
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>