[Autotest] - Add logging to check for telemetry pre/post test
Something somewhere is deleting this folder. I can't find *exactly*
where, and I have spent a ton of time digging into changes in both
autotest/ tast and nothing stands out as guilty.

At this point I am going to add this log, and hopefully it will allow me
to trace back to a test which has it prior, but not after (thus is the
guilty party)

BUG=b:203609358
TEST=dummy_Pass with/without the folder and verified the logs.

Change-Id: I792ad87c08847332f9c4a1491e2752103d24c267
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3235831
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Seewai Fu <seewaifu@google.com>