autoupdate_Basic: Add n2n and m2n versions that use a PIN login.
We recently added a PIN login test. We want to use that as part of the
important AU tests that do a login before and after the update.

This CL adds an n2n and m2n version of that test. However, the m2n is
not scheduled yet as we need to wait for M-96 to go to stable (since
that is the earlier milestone that has the PIN login test on the
branch).

BUG=b:197667947
TEST=login_LoginPin
TEST=autoupdate_Basic.m2n.full (still works)
TEST=autoupdate_Basic.full.pin (works)
TEST=autoupdate_Basic.m2n.full.pin (fails at first login)

Change-Id: Id944408ae43be7b0627cd3d463f4ea36038761a0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3238116
Commit-Queue: David Haddock <dhaddock@chromium.org>
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>