[autotest] Fix base_servohost.py's copy of _make_scp_cmd
Add logic to default the port of _make_scp_cmd in the class BaseServoHosst to 22.
In order to avoid disturbing the rest of the code, I am applying a default value
of 22 in this function alone and logging a message when the default is applied.

BUG=b/204502754
TEST=CQ

Change-Id: Ia1e8b7ff6634440db43712bb4d8acf1430b3459f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3251444
Tested-by: Gregory NISBET <gregorynisbet@google.com>
Reviewed-by: Xianuo Wang <xianuowang@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Gregory NISBET <gregorynisbet@google.com>