firmware_test: Support MiniOS kernel backup/restore
Extend the kernel utilities in firmware_test.py to support not only the
partitions 'KERN' but also 'MINIOS'. The extended methods are:
* is_kernel_changed
* backup_kernel
* is_kernel_saved
* restore_kernel

By default, theses methods will still backup or restore the partitions
with 'KERN-' and we can set kernel_type='MINIOS' to modify MiniOS
instead.

This change will not affect the original utilities of backup and restore
kernels. This tool will still generate the kernel_(A|B).original for
their backups and minios_(A|B) for MiniOS backups.

BUG=b:202217528
TEST=/usr/bin/test_that --fast --no-retry --iterations 1 --board=brya
     $DUT_IP --autotest_dir $AUTOTEST_DIR firmware_CorruptKernelA.dev
TEST=/usr/bin/test_that --fast --no-retry --iterations 1 --board=brya
     $DUT_IP --autotest_dir $AUTOTEST_DIR firmware_CorruptKernelB.dev
TEST=Test corrupt MiniOS with the following CLs

Change-Id: Iddc7168bb179ce2af130cc46fb128ac966a19c80
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3256630
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>