faft_checkers: Add checker dev_boot_minios_checker()
Add the checker of checking if DUT is in developer mode and booted to
MiniOS successfully.

MiniOS could be accessed by SSH if DUT is in developer mode, but there
is no autotest library in MiniOS, so we could only leverage the raw
command host.run_output to verify the boot.

Also rename the private attribute in RPCProxy:
* RPCProxy._client -> .host (also change it to public attribute)
so that the CrosHost instance could be accessed.

BUG=b:202217528
TEST=none

Cq-Depend: chromium:3220994
Change-Id: Iaff1279a9fda0b5b6e212cc9f022039beee68a02
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3259101
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>