server/site_tests: Add firmware_CorruptBothMiniosAB
Add a new test firmware_CorruptBothMiniosAB to test that DUT will fail
to boot MiniOS and stay in the recovery screen if both of the MiniOS
partitions are broken.

This test use dd to corrupt and restore the partitions, and leverages
USB boots to detect if DUT stayed in the recovery screen.

BUG=b:202217528
TEST=/usr/bin/test_that --fast --no-retry --iterations 1 --board=brya
     $DUT_IP --autotest_dir $AUTOTEST_DIR
     firmware_CorruptBothMiniosAB

Cq-Depend: chromium:3220994, chromium:3261998
Change-Id: I8fc631f50ff41ee0f911679d825511ca2931fe7f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3263675
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Commit-Queue: Hsuan Ting Chen <roccochen@chromium.org>