firmware_Cr50RddG3: modify errors
Rdd issues may not reliably show up on boards. The faft-config options
just note that Rdd may be unreliable on the board. Log a warning instead
of failing the test if the board is marked with a Rdd issue, but the
issue doesn't show up in a test run.

BUG=none
TEST=run on hatch

Change-Id: Ib425945d6cb7cc8d636e3852261ccd48d082612e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3260938
Tested-by: Mary Ruthven <mruthven@chromium.org>
Auto-Submit: Mary Ruthven <mruthven@chromium.org>
Reviewed-by: Aadit Maheshkumar Modi <modia@chromium.org>
Commit-Queue: Aadit Maheshkumar Modi <modia@chromium.org>