p2p: Re-enable p2p_EndToEndTest in the new skylab MultiDUT world.
I ended up gutting the test as it wanted to run over any arbitrary N
machines that were input. I have simplified it and just made it work
with a DUT and a companion DUT.

Delete this file that was used with it before that isn't used anywhere
else in autotest.

BUG=b:187795510
TEST=p2p_EndToEndTest against atlas|atlas deployment.

Change-Id: I9d98a204c8a16df0c13dbef440fc50f0ea2c5a5a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3266494
Tested-by: David Haddock <dhaddock@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Reviewed-by: Kyle Shimabukuro <kyleshima@chromium.org>