Add CPCon upload script to contrib, fix dependencies
Migrated the CPCon upload script from a non-main branch in
platform/moblab to the autotest contrib directory. Fix missing protobuf
information by changing the jobserializer to include a default board parameter in the
constructor, and allowing for the board parameter to be populated by the
board parsed from the model information in test results.

BUG=b:206797791
TEST=Local chroot

Change-Id: Iea25843e1d58521436b51a47b0fdd5b7d3dbd432
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3289983
Tested-by: Parker Holloway <jparkerh@google.com>
Reviewed-by: Svitlana Soltysyak <ssoltysyak@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Parker Holloway <jparkerh@google.com>