Bluetooth: Only test a2dp is connected if expecting a fail
After crrev/c/3245050, test_a2dp_sinewaves has no longer a single test
but a series of tests. This breaks the assumption in expect_pass since
it resets the expected_pass to True after every test.

This change makes EP tests only check if a2dp is connected on raspi if
we are expecting the test to fail. This is fine because:
- if a2dp is connected, that means service allowlist is not working as
  expected.
- if a2dp is not connected, it is pointless to continue the test because
  they'll certainly fail.

BUG=b:205618219
TEST=run bluetooth_AdapterEPHealth on hatch

Change-Id: Iba923cb063b9026a65e16b5739c7e3161cbf0df8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3288534
Tested-by: Yun-Hao Chung <howardchung@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>