firmware_ECPowerButton: Add the external display for chromebox
When running the firmware_ECPowerButton on the chromebox, the tester may
connect the external display to the device. Under this condition, the
device won't shutdown when pressing the power key with 200ms. To avoid
this failure, we add a check for the external display.

BUG=b:210083598, b:214114111
TEST=test_that firmware_ECPowerButton pass with/without external monitor

Change-Id: Ibf68f217c9bacade830e32756ab23f13df90607e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3299815
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Tested-by: Zhuohao Lee <zhuohao@chromium.org>
Auto-Submit: Zhuohao Lee <zhuohao@chromium.org>
Reviewed-by: Otabek Kasimov <otabek@google.com>
Commit-Queue: Zhuohao Lee <zhuohao@chromium.org>