bluetooth: Migrate _dbus_byte_array_to_b64_string to Python3.
Now the Bluetooth facade is run in Python 3.
_dbus_byte_array_to_b64_string() is used to generate results of XMLRPC
methods. In Python2, the encoded base64 bytes are treated as string and
can be used for XMLRPC result. However, in Python 3, bytes type is not
string type. So we have to decode() base64 encoded bytes to base64
string.

BUG=b:205356379
TEST=Run bluetooth_AdapterLEBetterTogether.smart_unlock_test test.

Change-Id: I79d35033c4b639a61216d79027691ee3523432eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3304713
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>
Tested-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Shuo-Peng Liao <deanliao@chromium.org>