autotest - utils shebangs to py3
Not including build_externals as that is a pure python2 op that needs to
be removed anyways. Its currently, or was last i checkd, tied to drone
builds, so I will not remove it in this CL.

Deleted modelviz as I can't find a call to it anywhere.

BUG=b:187788514
TEST=cq, unittest_suite, test_that

Change-Id: I2d74847832dbd9d6b4d50f3cc029e9e0ed37643c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3311900
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Commit-Queue: Derek Beckett <dbeckett@chromium.org>