Bluetooth: Always start discovery even if the device has been found
In tests that regarding multiple devices, we usually do:
    test_discover_device(peer_1)
    test_pairing(peer_1)
    (some tests to peer_1)
    test_discover_device(peer_2)
    test_pairing(peer_2)
    (some tests to peer_2)

In the following situation, test_pairing to peer_2 always fail:
peer_2 is found in the first discovery, and in the second discovery, DUT
still remember peer_2 so the second discovery doesn't be initiated. But
just before pairing with peer_2, bluez removes peer_2 because it is a
temporary device and hasn't been updated since the first discovery.

This CL resolves the above issue by removing the device object in
test_discover_device if the device has been found. This ensures we have
around 30 seconds to start pairing before bluez forgets the device.

BUG=b:207590909
TEST=run bluetooth_AdapterMDHealth for 5 iterations on hatch.
All 5 iterations are passed.

Change-Id: I984b98ebe9c1d38cb7c580636f643eea0225e711
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3319759
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Yun-Hao Chung <howardchung@chromium.org>
Tested-by: Yun-Hao Chung <howardchung@chromium.org>