Autotest: Output exception in perf ping test failure
The current code didn't include the exception, to improve debugging
we are including the exception in the error message.

The test fails as expected when running in pool without the necessary
pcap to router connection for traffic generation. After the change the
failure includes the exception.

AUTHOR=kglund
BUG=None
TEST=test_that --autotest_dir=$HOME/trunk/src/third_party/autotest/files/ --args="router_addr=$ROUTER pcap_addr=$PCAP" $DUT network_WiFi_Perf.ht20

Change-Id: Id27cef8042e453c47d69100785c0df773b232e8b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3321292
Reviewed-by: Joshua Emele <jemele@chromium.org>
Auto-Submit: Frank Gorgenyi <frankgor@chromium.org>
Commit-Queue: Frank Gorgenyi <frankgor@chromium.org>
Tested-by: Frank Gorgenyi <frankgor@chromium.org>