Disable Tauto policy test executions
To keep running these tests, a Python 3 migration would be necessary.
As there is an ongoing migration to Tast already, there is no gain to
migrate these test to Python 3 for a short time.

This is a fix for change I5118b7657ce0733343eef7629faf93295c3593ca,
which meant to disable these tests.

BUG=None
TEST=None

Change-Id: Ia500e4a389e5ec67901632f3988b64a731da1c12
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3322715
Tested-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Gabor Zsolt Magda <gabormagda@google.com>