fw testing: rm second and final obsolete PD test
This test ends up invoking ectool with the --name=cros_pd flag, meaning
it is not written to test the primary CrOS EC (--name=cros_ec).  cros_pd
controllers are not in the current support matrix, so this test looks
safe to delete as it does not operate on any current devices.

BUG=b:209630776
TEST=None

Change-Id: Idff8f6b720f432aed0f5134da068308a7b0706de
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3368963
Tested-by: Kevin Shelton <kmshelton@chromium.org>
Reviewed-by: Jeremy Bettis <jbettis@chromium.org>
Commit-Queue: Kevin Shelton <kmshelton@chromium.org>