kernel_FirmwareRequest: Drop
It's not ported to python3, it hasn't been running for a long time, and
it's not trivial (for me) to get working again.

It's also part of the kernel selftests suite, so it would make more
sense to get test coverage that way, rather than continuing to package
it individually for Tauto.

BUG=none
TEST=none

Change-Id: I29b38cf320b0875fe3b51529d9edfe1c062ada73
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3378966
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Tested-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>