power: video: Always activate the target tab
Sometimes the video is played in the background tab. This may lead to
spurious power consumption numbers on video testing.

To solve the issue, we don't need to create a new tab for testing since
Chrome always starts with an empty one.
In addition, we force activate the target tab right before starting the
video to ensure it's in the foreground.

BUG=b:214010309,b:213258078
TEST=test_that $IP power_VideoPlayback

Change-Id: Ic81917ddbfb483f29ab695fe8a04ec9e8fec144b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3380341
Tested-by: Fei Shao <fshao@chromium.org>
Auto-Submit: Fei Shao <fshao@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Fei Shao <fshao@chromium.org>