autotest: Add 11ax HE checks and bandwidth types to router and pcap
The modified classes are WiFiCellPerfTestBase, TestBase and WiFiClient.
And 11ax cap checks enable DUT to add 11ax HE cap to it's caps array
for both router and pcap router.

Also add 3 bandwidth types: HE40, HE80 and HE160 which are advertised
as HE caps in iw command output. HE80_80 has almost no industrial
adoption so it's unsupported in the test framework. They were added to
router before but not to the pcap. This CL adds them to pcap.

BUG=b:204593635, b:244645527
TEST=Ran on Ubiquiti UniFi 6 Lite AP with OpenWRT image, \
     test_that --board=volteer $DUT network_WiFi_Perf.ht20, and then \
     find logs showing DUT's capabilities: {'5ghz', 'he', ...}
     on volteer (AX201).
TEST=On gale, test_that --board=volteer $DUT network_WiFi_Perf.ht20
     ran successfully. logs could also show DUT's caps {'he', ...}.
TEST=test_that --board=volteer $DUT \
     network_WiFi_ChaosConnectDisconnect.static_5Ghz passes on gale

Change-Id: I6bf4a84374ab6477cc5d1a91894582880af255a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3418957
Reviewed-by: Seewai Fu <seewaifu@google.com>
Commit-Queue: Jun Yu <junyuu@chromium.org>
Reviewed-by: Billy Zhao <billyzhao@chromium.org>
Tested-by: Jun Yu <junyuu@chromium.org>