bluetooth: Specify Chameleon bundle per milestone
If we deploy a new Chameleon bundle with an incompatible interface
change, we need to back-port the Autotest tests changes into previous
milestones. However, back-port to stable branch is discouraged unless
absolutely necessary. Also, in the lab there are tests of M-k (k>=3)
milestone running occasionally, which is unlikely to be back-ported.
So here we propose an mechanism for developers to specify Chameleon
bundle version for each milestone.

We can maintain a table about the mappings of milestones to the
chameleon commits in a file, say, bluetooth_commits.yaml, in the Google
storage cloud. Use this table to check which chameleon bundle version
should be deployed to run the tests.

The new bluetooth_commits.yaml in the local is only for tracking history
of bluetooth_commits.yaml that in the cloud. Note that
NEEDS_SYSTEM_UPDATE flag is added in https://crrev.com/c/3462040.

BUG=b:214496518
TEST=test_that bluetooth_AdapterAUHealth.au_a2dp_test
TEST=test_that bluetooth_PeerUpdate

Change-Id: If2a0a5f66917a34312b6afb68e5969a45141379b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3386750
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Shyh-In Hwang <josephsih@chromium.org>
Commit-Queue: John Lai <johnlai@google.com>
Tested-by: John Lai <johnlai@google.com>