power: Run python3_convert on client/cros/power directory
This fixes an issue with power_BatteryCharge incorrectly finding that
the system is not plugged into AC due to an issue with
power_status.get_status. python3_convert fixes this. This also finishes
the conversion for the reset of the power/ directory.

BUG=b:203802087
TEST=test_that --board atlas --py_version=3 <dut> power_BatteryCharge

Change-Id: Ieb4a2dfcf4e7c99d4937c5a7d14a7072f2abdfca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3402497
Reviewed-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: Derek Basehore <dbasehore@chromium.org>
Tested-by: Derek Basehore <dbasehore@chromium.org>
Auto-Submit: Derek Basehore <dbasehore@chromium.org>