py3: test & reenable power_USBHotplugInSuspend
This test relied on a (broken) feature to run a test in the background.
To make it work again, this change introduces a `suspend_bg` function
into `cros_host` for the benefit of other users as well. It does the
exact same as the `suspend` function, but rather than wait, verify, and
return when the device resumes, it returns right away, allowing the
caller to perform work while the device is suspended. In our case said
work is plugging/unplugging a USB device.

BUG=b:203802087
TEST=export PY_VERSION=3
     test_that -b hatch 10.0.0.30 power_USBHotplugInSuspend
[...]
/tmp/test_that_results_hatch_Swlm0T/results-1-power_USBHotplugInSuspend
[  PASSED  ]
[...]
Total PASS: 2/2 (100%)

Change-Id: I240daad0435c4e3ab4c57fdb11bec48ebbdcd3d5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3403749
Tested-by: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Auto-Submit: Ruben Rodriguez Buchillon <coconutruben@chromium.org>
Reviewed-by: Derek Beckett <dbeckett@chromium.org>
Commit-Queue: Ruben Rodriguez Buchillon <coconutruben@chromium.org>