FAFT: Prevent unnecessary reboots in MiniOS related tests
In MiniOS related tests, we always perform a reboot as the first step of
cleanup(). This ensures that we could always boot back to the OS even if
the test failed during running it.

However, for the devices which was skipped and raised TestNA, it wastes
time for this unnecessary reboot and cleanup step. So we add an
additional check to prevent this in the cleanup() stage.

This applies for the following FAFT tests:
* firmware_MiniosPriority
* firmware_CorruptMinios
* firmware_CorruptBothMinios
* firmware_MiniosMenu

BUG=b:202217528
TEST=Add 'minios_enabled': false in fw-testing-configs/brya.json,
     run consolidate.py, run FAFT with the following command:
     /usr/bin/test_that --fast --no-retry --iterations 1 --board=brya
     $DUT_IP --autotest_dir $AUTOTEST_DIR
     firmware_MiniosPriority.minios_a firmware_MiniosPriority.minios_b
     firmware_CorruptMinios.minios_a firmware_CorruptMinios.minios_b
     firmware_CorruptBothMiniosAB
     firmware_MiniosMenu firmware_MiniosMenu.old
     and ensure that tests were skipped and DUT did not reboot

Change-Id: I76a20dd5bcb2af600cfdf182ce0aec41007049eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3411690
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>
Tested-by: Hsuan Ting Chen <roccochen@chromium.org>
Auto-Submit: Hsuan Ting Chen <roccochen@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>