Bluetooth: make untrusted pair for adapter connect_disconnect test
A BlueZ change wants to land to allow the trusted device to
auto-reconnect, which allows the UI to match the upstream
implementation. However, this will break the existing
connect_disconnect_loop test, assuming all auto-reconnect attempts are
blocked. This patch updates the test not to mark paired peer devices to
be trusted to continue to stop them from reconnecting.

BUG=b:209518368,b:216395140
TEST=bluetooth_AdapterLESanity.le_connect_disconnect_loop
TEST=bluetooth_AdapterCLHealth.cl_connect_disconnect_loop_test

Change-Id: I884f0f181d881ed25ce9edd1574da53f59b46c34
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3425239
Tested-by: Michael Sun <michaelfsun@google.com>
Reviewed-by: Alain Michaud <alainm@chromium.org>
Commit-Queue: Michael Sun <michaelfsun@google.com>