autotest - add workaround for telemetry six version
original idea in crrev.com/c/3413987 didn't work because it looks like
the import happens earlier in the stack. Forcing a reload. I tested this
in an SSP container on the drone by importing six, then common, then
running this mod. The version was 1.16... That said, the ssp environment
is a bit of magic man behind the curtain, and things happen in there I
just can't test locally, so I am not 100.00% sure this is the fix, as
there might be something in there messing up the location of the egg or
something I am unaware of. Hopefully this gets us there though.

BUG=b:206008069
TEST=cq

Change-Id: I09782dc03600695629fafe162133b3663c481e54
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/autotest/+/3429854
Tested-by: Derek Beckett <dbeckett@chromium.org>
Auto-Submit: Derek Beckett <dbeckett@chromium.org>
Reviewed-by: Jesse McGuire <jessemcguire@google.com>
Reviewed-by: Denis Nikitin <denik@chromium.org>
Commit-Queue: Denis Nikitin <denik@chromium.org>